+++ CHANGELOG.md
@@ -1,1 +1,11 @@
-Initial release
+Ok so this is the first actual release.
+
+Everything is very WIP and I'll fix bugs as I find them.
+
+# Changelog
+
+## New features:
+- added the color syntax specifically for infill
+- added the highlight syntax specifically for infill
+- added the underline syntax specifically for infill
+